% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O_E.CUSUM.R
\name{O_E.CUSUM}
\alias{O_E.CUSUM}
\title{plots the O-E CUSUM for each facility/center strata}
\usage{
O_E.CUSUM(
  date,
  ESCO_ord,
  data_path,
  facility_id_name,
  start_name,
  end_name,
  dial_drd_name,
  xbeta_name,
  patient_id_name,
  start_date,
  end_date,
  HPT1,
  HPT2,
  Lambda_list_prior,
  year_event_rate,
  pre,
  ESCO_de_identified_path,
  y_axis_path,
  h_val_path,
  rho
)
}
\arguments{
\item{date}{The day you use the function}

\item{ESCO_ord}{Grouping strata of the center/facility}

\item{data_path}{The path you saved the data}

\item{facility_id_name}{The column name of facility id}

\item{start_name}{The column name of entering follow up for each subject}

\item{end_name}{The column name of ending follow up for each subject}

\item{dial_drd_name}{The column name of qualifying deaths}

\item{xbeta_name}{The column name of adjusted risk}

\item{patient_id_name}{The column name of subject id}

\item{start_date}{The start date of the follow up (calendar time)}

\item{end_date}{The end date of the follow up (calendar time)}

\item{HPT1}{The H_1 of deterioration}

\item{HPT2}{The H_1 of improvement}

\item{Lambda_list_prior}{The unadjusted national average hazard rate for each subject i}

\item{year_event_rate}{The risk}

\item{pre}{Name of the center/facility}

\item{ESCO_de_identified_path}{Path of facility/center strata}

\item{y_axis_path}{Path of y_axis}

\item{h_val_path}{Path of control limits for different sample sizes}

\item{rho}{Parameter to control the restart}
}
\value{
the signal points and not signal points
}
\description{
The function takes the information for survival analysis and returns the signal situation and plots
}
\examples{
Mortality_CUSUM(date,ESCO_ord,"data_all.csv","facility_id_name","start_name","end_name",
"dial_drd_name","xbeta_name","patient_id_name",start_date,end_date,
HPT1,HPT2,Lambda_list_prior,year_event_rate,pre,
ESCO_de_identified_path,y_axis_path,h_val_path,rho)
#Output: graphs
}
